# docs/swagger.yaml
basePath: /api
definitions:
  auth.AuthResponse:
    properties:
      role:
        type: string
      token:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        maxLength: 60
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.SignupRequest:
    properties:
      bar_number:
        type: string
      email:
        maxLength: 120
        type: string
      jurisdiction:
        description: Optional for lawyers
        type: string
      name:
        maxLength: 80
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 6
        type: string
      role:
        enum:
        - client
        - lawyer
        type: string
    required:
    - email
    - name
    - password
    - role
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: machine-friendly error code
        example: FORBIDDEN
        type: string
      error:
        description: always true for error cases
        example: true
        type: boolean
      message:
        description: human-readable error message
        example: Forbidden
        type: string
    type: object
  models.Role:
    enum:
    - client
    - lawyer
    type: string
    x-enum-varnames:
    - RoleClient
    - RoleLawyer
  models.User:
    properties:
      barNumber:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      jurisdiction:
        type: string
      name:
        type: string
      passwordHash:
        type: string
      role:
        $ref: '#/definitions/models.Role'
    type: object
  models.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        description: field name -> list of error messages
        type: object
      message:
        example: Validation failed
        type: string
    type: object
...
info:
  contact:
    email: aldoetobex@gmail.com
    name: Aldo Rifki Putra
  description: 'API for a mini legal marketplace: clients post cases, lawyers submit
    quotes, clients accept & pay, and lawyers access files via signed URLs.'
  title: Mini Legal Marketplace API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate and receive a JWT
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /me:
    get:
      description: Return full profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
...
  /signup:
    post:
      consumes:
      - application/json
      description: Register a new user (client or lawyer)
      parameters:
      - description: Signup payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ValidationErrorResponse'
        "409":
          description: email already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign up
      tags:
      - auth
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"